requirements.txt
opencv-contrib-python==4.4.0.46
opencv-python==4.2.0.32
torch==1.11.0
torchvision==0.12.0+cu102
Pillow==9.0.1
pandas==1.1.5
scikit-image==0.15.0
scikit-learn==0.24.0
